+++ src/main/java/com/travelock/server/dto/block/BigBlockResponseDTO.java
@@ -1,14 +1,16 @@
-package com.travelock.server.dto;
+package com.travelock.server.dto.block;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class StateResponseDTO {
-    private Long stateId;
-    private String stateCode;
+public class BigBlockResponseDTO {
+    private Long bigBlockId;
+    private String cityCode;
+    private String cityName;
     private String stateName;
 }
