+++ src/main/java/com/travelock/server/dto/block/StateResponseDTO.java
@@ -1,16 +1,14 @@
-package com.travelock.server.dto;
+package com.travelock.server.dto.block;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class BigBlockResponseDTO {
-    private Long bigBlockId;
-    private String cityCode;
-    private String cityName;
+public class StateResponseDTO {
+    private Long stateId;
+    private String stateCode;
     private String stateName;
 }
