+++ src/main/java/com/travelock/server/service/BigBlockService.java
@@ -7,11 +7,9 @@ import com.travelock.server.domain.BigBlock;
 import com.travelock.server.domain.QBigBlock;
 import com.travelock.server.domain.QState;
 import com.travelock.server.domain.State;
-import com.travelock.server.dto.BigBlockResponseDTO;
-import com.travelock.server.dto.StateResponseDTO;
+import com.travelock.server.dto.block.BigBlockResponseDTO;
+import com.travelock.server.dto.block.StateResponseDTO;
 import com.travelock.server.exception.base_exceptions.ResourceNotFoundException;
-import com.travelock.server.repository.BigBlockRepository;
-import com.travelock.server.repository.StateRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
